+++ src/main/java/brawljars/api/intern/players/battlelog/BattleLogRequest.java
@@ -43,10 +43,4 @@ public class BattleLogRequest extends PaginationRequest {
         map.put("playerTag", playerTag);
         return map;
     }
-
-    @Override
-    public Map<String, Object> getQueryParameters() {
-        Map<String, Object> map = super.getQueryParameters();
-        return map;
-    }
 }