+++ src/main/java/brawljars/api/intern/rankings/powerplay/PowerplayRankingsRequest.java
@@ -51,10 +51,4 @@ public class PowerplayRankingsRequest extends PaginationRequest {
         map.put("seasonId", seasonId);
         return map;
     }
-
-    @Override
-    public Map<String, Object> getQueryParameters() {
-        Map<String, Object> map = super.getQueryParameters();
-        return map;
-    }
 }