+++ src/test/java/brawljars/api/BaseApiTest.java
@@ -17,7 +17,6 @@
 package brawljars.api;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.argThat;
@@ -55,24 +54,19 @@ class BaseApiTest {
 
     @Mock private Connector connector;
 
-    private ApiContext apiContext;
-
     private PaginationRequest request;
 
     private static ArgumentMatcher<RequestContext> createRequestContextArgumentMatcher(
             RequestContext expected) {
         return actual -> {
-            assertNotNull(actual);
-            assertEquals(expected.getApiKey(), actual.getApiKey());
-            assertEquals(expected.getUrl(), actual.getUrl());
-            assertEquals(expected.getResponseClass(), actual.getResponseClass());
+            assertEquals(expected, actual);
             return true;
         };
     }
 
     @BeforeEach
     void setUp() {
-        apiContext = new ApiContext(URL, API_KEY, connector);
+        ApiContext apiContext = new ApiContext(URL, API_KEY, connector);
         request = new FooRequest(100, "after", "before", true);
         unitUnderTest = new BaseApi(apiContext);
     }
@@ -129,7 +123,7 @@ class BaseApiTest {
     void get_whenWithValidParameters_shouldReturnResponse() throws Exception {
         FooResponse expected = new FooResponse();
         RequestContext requestContext =
-                new RequestContext("urlpart", API_KEY, request, FooResponse.class);
+                new RequestContext(URL + PART, API_KEY, request, FooResponse.class);
         when(connector.get(argThat(createRequestContextArgumentMatcher(requestContext))))
                 .thenReturn(expected);