+++ CHANGELOG.md
@@ -2,11 +2,31 @@
 
 ## [Unreleased]
 
+## [3.0.3] - 2023-09-23
+
+### Added
+
+- Usage of Java 17 HTTP Client
+
+### Fixed
+
+- Some minor code cleaning regarding Java 17 features
+
+### Changed
+
+- ApiContext to be a record
+
+### Removed
+
+- Apache HTTP Client dependency
+- Some redundant overridden methods in request objects
+
 ## [3.0.2] - 2023-07-12
 
 ### Added
 
 - Create release job #41
 
-[unreleased]: https://github.com/mlieshoff/brawljars/compare/v3.0.2...HEAD
+[unreleased]: https://github.com/mlieshoff/brawljars/compare/v3.0.3...HEAD
+[3.0.3]: https://github.com/mlieshoff/brawljars/compare/v3.0.2...v3.0.3
 [3.0.2]: https://github.com/mlieshoff/brawljars/compare/v3.0.2...v3.0.2